requirements.txt
basenet-api>=1.4.6
tensorflow~=2.10.0
tensorflow-gpu==2.10.*
keras == 2.10.0
graphviz >= 0.20.1
matplotlib==3.6.1
future~=0.18.2
Pillow~=9.2.0
numpy==1.23.4
ttkwidgets~=0.12.1
pydot~=1.4.2
# Graphviz.
# CUDA.
# CuDNN.
# pdflatex.
setuptools~=60.2.0